+++ docs/blob/azureExplorer.md
@@ -5,4 +5,4 @@ To see what is inside your blob, you can use Azure explorer running on
 
 ## Known issues with third party image
 This tool looks cool and is fully configured, but there is no working support for 
-deleting and downloading files, tested with pd
+deleting and downloading files, tested with pdf/png
