+++ src/FluentTesting.Common/Extensions/ApplicationFactoryBuilderExtensions.cs
@@ -4,19 +4,19 @@ using System.Diagnostics;
 
 namespace FluentTesting.Common.Extensions
 {
-    public static class ApplicationFactoryBuilderExtensions
-    {
-        /// <summary>
-        /// Not use proxied images - use with caution on local machine in case of lack of VPN
-        /// </summary>
-        public static IApplicationFactoryBuilder UseProxiedImages(this IApplicationFactoryBuilder builder, string proxy)
-        {
-            Trace.WriteLine("DO NOT FORGET TO REMOVE THIS BEFORE PUSHING TO AzDo!!");
+	public static class ApplicationFactoryBuilderExtensions
+	{
+		/// <summary>
+		/// Not use proxied images - use with caution on local machine in case of lack of VPN
+		/// </summary>
+		public static IApplicationFactoryBuilder UseProxiedImages(this IApplicationFactoryBuilder builder, string proxy)
+		{
+			Trace.WriteLine("DO NOT FORGET TO REMOVE THIS BEFORE PUSHING TO AzDo!!");
 
-            ImagesSetting.ImageProxyPath = proxy;
+			ImagesSetting.ImageProxyPath = proxy;
 
-            builder.UseProxiedImages = true;
-            return builder;
-        }
-    }
+			builder.UseProxiedImages = true;
+			return builder;
+		}
+	}
 }
