+++ tests/Samples.AspApp.Tests/Endpoints/AssertAuthJson.json
